Move `/helper` directly into `/src`
