+++ src/scripts/comic/getHoverChangerRelative.ts
@@ -1,7 +1,7 @@
-import setValue from '@/renderengine/helper/setValue'
+import setValue from '@/helper/setValue'
 
-import type { DoHover } from '@/renderengine/helper/typeHover'
-import type { SizeHover } from '@/renderengine/helper/typeSize'
+import type { DoHover } from '@/helper/typeHover'
+import type { SizeHover } from '@/helper/typeSize'
 
 const getHoverChangerRelative = (): {
   doHover: DoHover
