+++ src/scripts/persons-and-trees/builder.js
@@ -1,5 +1,5 @@
-import getHoverChangerStandard from '@/renderengine/helper/getHoverChangerStandard'
-import { getRandom } from '@/renderengine/helper/helper'
+import getHoverChangerStandard from '@/helper/getHoverChangerStandard'
+import { getRandom } from '@/helper/helper'
 
 import { Object } from './object.js'
 import { Arm, Shield, ShoulderPad, Sword, Tool } from './person-arm.js'
